+++ simple_history/tests/models.py
@@ -25,6 +25,14 @@ class Poll(models.Model):
         return reverse("poll-detail", kwargs={"pk": self.pk})
 
 
+class PollWithNonEditableField(models.Model):
+    question = models.CharField(max_length=200)
+    pub_date = models.DateTimeField("date published")
+    modified = models.DateTimeField(auto_now=True, editable=False)
+
+    history = HistoricalRecords()
+
+
 class PollWithUniqueQuestion(models.Model):
     question = models.CharField(max_length=200, unique=True)
     pub_date = models.DateTimeField("date published")
@@ -69,7 +77,7 @@ class PollWithExcludedFKField(models.Model):
 
 class AlternativePollManager(models.Manager):
     def get_queryset(self):
-        return super(AlternativePollManager, self).get_queryset().exclude(id=1)
+        return super().get_queryset().exclude(id=1)
 
 
 class PollWithAlternativeManager(models.Model):
@@ -102,13 +110,13 @@ class PollWithHistoricalIPAddress(models.Model):
 class CustomAttrNameForeignKey(models.ForeignKey):
     def __init__(self, *args, **kwargs):
         self.attr_name = kwargs.pop("attr_name", None)
-        super(CustomAttrNameForeignKey, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
 
     def get_attname(self):
-        return self.attr_name or super(CustomAttrNameForeignKey, self).get_attname()
+        return self.attr_name or super().get_attname()
 
     def deconstruct(self):
-        name, path, args, kwargs = super(CustomAttrNameForeignKey, self).deconstruct()
+        name, path, args, kwargs = super().deconstruct()
         if self.attr_name:
             kwargs["attr_name"] = self.attr_name
         return name, path, args, kwargs
@@ -122,15 +130,13 @@ class ModelWithCustomAttrForeignKey(models.Model):
 class CustomAttrNameOneToOneField(models.OneToOneField):
     def __init__(self, *args, **kwargs):
         self.attr_name = kwargs.pop("attr_name", None)
-        super(CustomAttrNameOneToOneField, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
 
     def get_attname(self):
-        return self.attr_name or super(CustomAttrNameOneToOneField, self).get_attname()
+        return self.attr_name or super().get_attname()
 
     def deconstruct(self):
-        name, path, args, kwargs = super(
-            CustomAttrNameOneToOneField, self
-        ).deconstruct()
+        name, path, args, kwargs = super().deconstruct()
         if self.attr_name:
             kwargs["attr_name"] = self.attr_name
         return name, path, args, kwargs
@@ -138,7 +144,7 @@ class CustomAttrNameOneToOneField(models.OneToOneField):
 
 class ModelWithCustomAttrOneToOneField(models.Model):
     poll = CustomAttrNameOneToOneField(Poll, models.CASCADE, attr_name="custom_poll")
-    history = HistoricalRecords(excluded_field_kwargs={"poll": set(["attr_name"])})
+    history = HistoricalRecords(excluded_field_kwargs={"poll": {"attr_name"}})
 
 
 class Temperature(models.Model):
@@ -189,15 +195,13 @@ class Voter(models.Model):
 class HistoricalRecordsVerbose(HistoricalRecords):
     def get_extra_fields(self, model, fields):
         def verbose_str(self):
-            return "%s changed by %s as of %s" % (
+            return "{} changed by {} as of {}".format(
                 self.history_object,
                 self.history_user,
                 self.history_date,
             )
 
-        extra_fields = super(HistoricalRecordsVerbose, self).get_extra_fields(
-            model, fields
-        )
+        extra_fields = super().get_extra_fields(model, fields)
         extra_fields["__str__"] = verbose_str
         return extra_fields
 
@@ -224,7 +228,7 @@ class Person(models.Model):
         if hasattr(self, "skip_history_when_saving"):
             raise RuntimeError("error while saving")
         else:
-            super(Person, self).save(*args, **kwargs)
+            super().save(*args, **kwargs)
 
 
 class FileModel(models.Model):
@@ -251,6 +255,7 @@ class Document(models.Model):
     changed_by = models.ForeignKey(
         User, on_delete=models.CASCADE, null=True, blank=True
     )
+
     history = HistoricalRecords()
 
     @property
@@ -269,6 +274,12 @@ class Paper(Document):
         self.changed_by = value
 
 
+class RankedDocument(Document):
+    rank = models.IntegerField(default=50)
+
+    history = HistoricalRecords()
+
+
 class Profile(User):
     date_of_birth = models.DateField()
 
@@ -284,7 +295,9 @@ class State(models.Model):
 
 class Book(models.Model):
     isbn = models.CharField(max_length=15, primary_key=True)
-    history = HistoricalRecords(verbose_name="dead trees")
+    history = HistoricalRecords(
+        verbose_name="dead trees", verbose_name_plural="dead trees plural"
+    )
 
 
 class HardbackBook(Book):
@@ -301,6 +314,7 @@ class Library(models.Model):
 
     class Meta:
         verbose_name = "quiet please"
+        verbose_name_plural = "quiet please plural"
 
 
 class BaseModel(models.Model):
@@ -368,6 +382,14 @@ class UnicodeVerboseName(models.Model):
 
     class Meta:
         verbose_name = "\u570b"
+
+
+class UnicodeVerboseNamePlural(models.Model):
+    name = models.CharField(max_length=100)
+    history = HistoricalRecords()
+
+    class Meta:
+        verbose_name_plural = "\u570b"
 
 
 class CustomFKError(models.Model):
@@ -683,13 +705,11 @@ class OverrideModelNameAsString(models.Model):
 
 class OverrideModelNameAsCallable(models.Model):
     name = models.CharField(max_length=15, unique=True)
-    history = HistoricalRecords(custom_model_name=lambda x: "Audit{}".format(x))
+    history = HistoricalRecords(custom_model_name=lambda x: f"Audit{x}")
 
 
 class AbstractModelCallable1(models.Model):
-    history = HistoricalRecords(
-        inherit=True, custom_model_name=lambda x: "Audit{}".format(x)
-    )
+    history = HistoricalRecords(inherit=True, custom_model_name=lambda x: f"Audit{x}")
 
     class Meta:
         abstract = True