+++ simple_history/tests/tests/utils.py
@@ -12,7 +12,7 @@ middleware_override_settings = {
 }
 
 
-class TestDbRouter(object):
+class TestDbRouter:
     def db_for_read(self, model, **hints):
         if model._meta.app_label == "external":
             return OTHER_DB_NAME
@@ -42,7 +42,7 @@ database_router_override_settings = {
 }
 
 
-class TestModelWithHistoryInDifferentDbRouter(object):
+class TestModelWithHistoryInDifferentDbRouter:
     def db_for_read(self, model, **hints):
         if model == HistoricalModelWithHistoryInDifferentDb:
             return OTHER_DB_NAME